backend/requirements_minimal.txt
# Core Django
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Databases
psycopg2-binary
pymongo
pgvector

# File handling
python-magic
Pillow
PyPDF2
python-docx
python-pptx
openpyxl
beautifulsoup4

# AI/ML
requests

# Utilities
python-dotenv
jsonschema
